Move computation around
