+++ src/fastxsketch.h
@@ -43,6 +43,7 @@ using FastxSketchingResult = SketchingResult;
 
 FastxSketchingResult fastx2sketch(Dashing2Options &opts, const std::vector<std::string> &paths);
 FastxSketchingResult fastx2sketch_byseq(Dashing2Options &opts, const std::string &path, kseq_t *kseqs=static_cast<kseq_t *>(nullptr), bool parallel=false, const size_t seqs_per_batch = 8192);
+std::string makedest(Dashing2Options &opts, const std::string &path, bool iskmer=false);
 }
 
 #endif
